6-size.c

+++ 0x00-hello_world/6-size.c
@@ -10,10 +10,10 @@ int main(void)
 	long int c;
 	long long int d;
 	float e;
-printf("Size of a char %lu byte(S)\n", (unsigned long)sizeof(a));
-printf("Size of an int %lu byte(S)\n", (unsigned long)sizeof(b));
-printf("Size of a long int %lu byte(S)\n", (unsigned long)sizeof(c));
-printf("Size of a long long int %lu byte(S)\n", (unsigned long)sizeof(d));
-printf("Size of a float %lu byte(S)\n", (unsigned long)sizeof(e));
+printf("Size of a char %zu byte(S)\n", (unsigned long)sizeof(a));
+printf("Size of an int %zu byte(S)\n", (unsigned long)sizeof(b));
+printf("Size of a long int %zu byte(S)\n", (unsigned long)sizeof(c));
+printf("Size of a long long int %zu byte(S)\n", (unsigned long)sizeof(d));
+printf("Size of a float %zu byte(S)\n", sizeof(e));
 return (0);
 }
